v1.2.12
Corrections to bitsetc allocator-awareness and consistency. Removal of some unnecessary pre-c++11 checks in bitsetb and bitset. Bitsetb now no longer sets the buffer to zero on default construction, as this interferes with correct allocator transfer with copy constructor in bitsetc. When using bitsetb directly, if the buffer is not already zero'ed when supplied to bitsetb, use bitsetb::reset() to do so.